Demangle symbols while processing the name section
From the wasm spec ([name section] and [name encoding]) symbol names
are encoded in utf8 so they do not need to be mangled in the final
artifact.  In addition to being easier to read, the demangled symbols
save a few additional bytes in the output.

Per the spec the assigned names do not need to be unique, however as
the default option the hash at the end of the symbol is left intact.
A command line argument could be added to customize this behavior in
the future.

The primary reason for leaving the hash is that we are not certain the
tool ecosystem will handle this in a consistent way.  For instance, it
would be confusing if `wat2wasm` adds a `.3` suffix to the 4th copy of
a function (0 indexed) while another tool adds `.4` resulting in
potential confusion.

[name section]: https://webassembly.github.io/spec/appendix/custom.html#name-section
[name encoding]: https://webassembly.github.io/spec/binary/values.html#names