Allow reusing the internal `parity_wasm::Module`
This commit allows to avoid parsing and serializing a wasm module, if possible.
This is one of the hot spots of `wasm-bindgen` right now, and it's already got
the same version of `parity_wasm` so there's no need to do this reserialization
and reparsing!

This is done in a fashion, though, to avoid making `parity_wasm` a public
dependency which is changing rapidly. Instead it uses `Any` to allow other
crates to pass in an arbitrary type, but otherwise a fallback for going through
bytes is always provided and possible.